Moving converters out. Fix broken tab from converter having null values
